fix: FileCommandService, FileQueryService interface 분리 및 중복된 파일 확장자 검증 로직 제거 및 최대 파일 크기 500MB로 증가
